+++ src/structlog/_loggers.py
@@ -7,18 +7,20 @@
 Logger wrapper and helper class.
 """
 
+from __future__ import annotations
+
 import copy
 import sys
 import threading
 
 from pickle import PicklingError
 from sys import stderr, stdout
-from typing import IO, Any, BinaryIO, Dict, Optional, TextIO
+from typing import IO, Any, BinaryIO, TextIO
 
 from structlog._utils import until_not_interrupted
 
 
-WRITE_LOCKS: Dict[IO[Any], threading.Lock] = {}
+WRITE_LOCKS: dict[IO[Any], threading.Lock] = {}
 
 
 def _get_lock_for_file(file: IO[Any]) -> threading.Lock:
@@ -53,7 +55,7 @@ class PrintLogger:
        monkeypatchability.
     """
 
-    def __init__(self, file: Optional[TextIO] = None):
+    def __init__(self, file: TextIO | None = None):
         self._file = file or stdout
 
         self._lock = _get_lock_for_file(self._file)
@@ -83,7 +85,7 @@ class PrintLogger:
 
         self._lock = _get_lock_for_file(self._file)
 
-    def __deepcopy__(self, memodict: Dict[Any, Any] = {}) -> "PrintLogger":
+    def __deepcopy__(self, memodict: dict[Any, Any] = {}) -> PrintLogger:
         """
         Create a new PrintLogger with the same attributes. Similar to pickling.
         """
@@ -127,11 +129,117 @@ class PrintLoggerFactory:
     .. versionadded:: 0.4.0
     """
 
-    def __init__(self, file: Optional[TextIO] = None):
+    def __init__(self, file: TextIO | None = None):
         self._file = file
 
     def __call__(self, *args: Any) -> PrintLogger:
         return PrintLogger(self._file)
+
+
+class WriteLogger:
+    """
+    Write events into a file.
+
+    :param file: File to print to. (default: `sys.stdout`)
+
+    >>> from structlog import WriteLogger
+    >>> WriteLogger().msg("hello")
+    hello
+
+    Useful if you follow
+    `current logging best practices <logging-best-practices>`.
+
+    Also very useful for testing and examples since logging is finicky in
+    doctests.
+
+    A little faster and a little less versatile than the PrintLogger.
+
+    .. versionadded:: 22.1
+    """
+
+    def __init__(self, file: TextIO | None = None):
+        self._file = file or sys.stdout
+        self._write = self._file.write
+        self._flush = self._file.flush
+
+        self._lock = _get_lock_for_file(self._file)
+
+    def __getstate__(self) -> str:
+        """
+        Our __getattr__ magic makes this necessary.
+        """
+        if self._file is stdout:
+            return "stdout"
+
+        elif self._file is stderr:
+            return "stderr"
+
+        raise PicklingError(
+            "Only WriteLoggers to sys.stdout and sys.stderr can be pickled."
+        )
+
+    def __setstate__(self, state: Any) -> None:
+        """
+        Our __getattr__ magic makes this necessary.
+        """
+        if state == "stdout":
+            self._file = stdout
+        else:
+            self._file = stderr
+
+        self._lock = _get_lock_for_file(self._file)
+
+    def __deepcopy__(self, memodict: dict[Any, Any] = {}) -> WriteLogger:
+        """
+        Create a new WriteLogger with the same attributes. Similar to pickling.
+        """
+        if self._file not in (sys.stdout, sys.stderr):
+            raise copy.error(
+                "Only WriteLoggers to sys.stdout and sys.stderr "
+                "can be deepcopied."
+            )
+
+        newself = self.__class__(self._file)
+
+        newself._write = newself._file.write
+        newself._flush = newself._file.flush
+        newself._lock = _get_lock_for_file(newself._file)
+
+        return newself
+
+    def __repr__(self) -> str:
+        return f"<WriteLogger(file={self._file!r})>"
+
+    def msg(self, message: str) -> None:
+        """
+        Write and flush *message*.
+        """
+        with self._lock:
+            until_not_interrupted(self._write, message + "\n")
+            until_not_interrupted(self._flush)
+
+    log = debug = info = warn = warning = msg
+    fatal = failure = err = error = critical = exception = msg
+
+
+class WriteLoggerFactory:
+    r"""
+    Produce `WriteLogger`\ s.
+
+    To be used with `structlog.configure`\ 's ``logger_factory``.
+
+    :param file: File to print to. (default: `sys.stdout`)
+
+    Positional arguments are silently ignored.
+
+    .. versionadded:: 22.1
+    """
+
+    def __init__(self, file: TextIO | None = None):
+        self._file = file
+
+    def __call__(self, *args: Any) -> WriteLogger:
+        return WriteLogger(self._file)
 
 
 class BytesLogger:
@@ -149,7 +257,7 @@ class BytesLogger:
     """
     __slots__ = ("_file", "_write", "_flush", "_lock")
 
-    def __init__(self, file: Optional[BinaryIO] = None):
+    def __init__(self, file: BinaryIO | None = None):
         self._file = file or sys.stdout.buffer
         self._write = self._file.write
         self._flush = self._file.flush
@@ -183,7 +291,7 @@ class BytesLogger:
         self._flush = self._file.flush
         self._lock = _get_lock_for_file(self._file)
 
-    def __deepcopy__(self, memodict: Dict[Any, Any] = {}) -> "BytesLogger":
+    def __deepcopy__(self, memodict: dict[Any, Any] = {}) -> BytesLogger:
         """
         Create a new BytesLogger with the same attributes. Similar to pickling.
         """
@@ -230,7 +338,7 @@ class BytesLoggerFactory:
     """
     __slots__ = ("_file",)
 
-    def __init__(self, file: Optional[BinaryIO] = None):
+    def __init__(self, file: BinaryIO | None = None):
         self._file = file
 
     def __call__(self, *args: Any) -> BytesLogger: